requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sentence-transformers
faiss-cpu
pymupdf
pdfplumber
pillow
transformers
torch
openai
python-multipart